refactor: move importer to internal package
