+++ internal/importer/importer_reader.go
@@ -1,4 +1,4 @@
-package main
+package importer
 
 import (
 	"encoding/csv"
