+++ internal/importer/importer_reader_test.go
@@ -1,4 +1,4 @@
-package main
+package importer
 
 import (
 	"encoding/json"
