+++ internal/importer/importer_wizard.go
@@ -1,4 +1,4 @@
-package main
+package importer
 
 import (
 	"fmt"
