+++ internal/importer/importer_wizard_file.go
@@ -1,4 +1,4 @@
-package main
+package importer
 
 import (
 	"strings"
