+++ internal/importer/importer_wizard_publish.go
@@ -1,4 +1,4 @@
-package main
+package importer
 
 import (
 	"fmt"
