+++ internal/importer/importer_wizard_review.go
@@ -1,4 +1,4 @@
-package main
+package importer
 
 import (
 	"fmt"
