+++ internal/importer/importer_wizard_template.go
@@ -1,4 +1,4 @@
-package main
+package importer
 
 import (
 	"strings"
