+++ internal/importer/importer_wizard_test.go
@@ -1,4 +1,4 @@
-package main
+package importer
 
 import (
 	"errors"
