+++ update_importer.go
@@ -2,6 +2,8 @@ package main
 
 import (
 	tea "github.com/charmbracelet/bubbletea"
+
+	"github.com/marang/emqutiti/internal/importer"
 )
 
 // updateImporter forwards messages to the import wizard when active.
@@ -10,7 +12,7 @@ func (m *model) updateImporter(msg tea.Msg) tea.Cmd {
 		return nil
 	}
 	nm, cmd := m.importWizard.Update(msg)
-	if w, ok := nm.(*ImportWizard); ok {
+	if w, ok := nm.(*importer.ImportWizard); ok {
 		m.importWizard = w
 	}
 	return cmd
